refactor(Frame): make abstract class for remove duplicate
