+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,17 +1,42 @@
 package bowling.domain.frame;
 
 import bowling.domain.score.Score;
+import bowling.domain.score.ScoreType;
+import bowling.domain.score.Scores;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public interface Frame {
-    void addScore(int point);
+public abstract class Frame {
+    private static final int ZERO_POINT = 0;
+    private static final int SECOND_PLAY = 1;
+    private static final int STRIKE_POINT = 10;
 
-    boolean isPlayable();
+    protected final Scores scores;
 
-    boolean isCalculatableFrame(int frameIndex);
+    public Frame(Scores scores) {
+        this.scores = scores;
+    }
 
-    List<Score> getScores();
+    public abstract void addScore(int point);
 
-    int getTotalPoint(int frameIndex);
+    public abstract boolean isPlayable();
+
+    public abstract boolean isCalculatableFrame(int frameIndex);
+
+    public abstract int getTotalPoint(int frameIndex);
+
+    public List<Score> getScores() {
+        return new ArrayList<>(scores.getScores());
+    }
+
+    protected ScoreType getSpareOrGutterType(Scores scores, int point) {
+        if (point == ZERO_POINT) {
+            return ScoreType.GUTTER;
+        }
+        if (scores.size() == SECOND_PLAY && scores.currentPoint() + point == STRIKE_POINT) {
+            return ScoreType.SPARE;
+        }
+        return ScoreType.MISS;
+    }
 }
