+++ src/main/java/bowling/domain/frame/LastFrame.java
@@ -13,21 +13,19 @@ import java.util.List;
  * 마지막 프레임을 나타내기 위한 객체
  * 마지막 프레임에서 스트라이크, 스페어 일 경우 1구를 더 진행한다.
  */
-public class LastFrame implements Frame {
+public class LastFrame extends Frame {
     private static final int STRIKE_POINT = 10;
     private static final int DEFAULT_PLAY_COUNT = 2;
     private static final int BONUS_PLAY_COUNT = 3;
     private static final int FIRST_PLAY = 0;
     private static final int SECOND_PLAY = 1;
     private static final int DOUBLE_STRIKE_POINT = 20;
-    private static final int ZERO_POINT = 0;
     private static final int THIRD_PLAY = 2;
 
-    private final Scores scores;
     private final BonusScores bonusScores;
 
     public LastFrame(BonusScores bonusScores) {
-        this.scores = new Scores();
+        super(new Scores());
         this.bonusScores = bonusScores;
     }
 
@@ -66,16 +64,6 @@ public class LastFrame implements Frame {
         return getSpareOrGutterType(scores, point);
     }
 
-    private ScoreType getSpareOrGutterType(Scores scores, int point) {
-        if (point == ZERO_POINT) {
-            return ScoreType.GUTTER;
-        }
-        if (scores.size() == SECOND_PLAY && scores.currentPoint() + point == STRIKE_POINT) {
-            return ScoreType.SPARE;
-        }
-        return ScoreType.MISS;
-    }
-
     private void validateSecondPoint(int point) {
         if (!scores.isStrike(FIRST_PLAY) && totalPoint(point) > STRIKE_POINT) {
             throw new IllegalArgumentException("마지막 포인트합은 10점을 넘을수 없습니다.");
@@ -102,11 +90,6 @@ public class LastFrame implements Frame {
         }
 
         return scores.size() < BONUS_PLAY_COUNT && scores.hasStrikeOrSpare();
-    }
-
-    @Override
-    public List<Score> getScores() {
-        return new ArrayList<>(scores.getScores());
     }
 
     @Override
